refactor: move app-sidebar to components folder
