+++ src/app/properties/[id]/layout.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "~/app/_components/app-sidebar";
+import { AppSidebar } from "~/components/ui/app-sidebar";
 import { SidebarProvider, SidebarTrigger } from "~/components/ui/sidebar";
 
 type Params = Promise<{ id: string }>;
